requirements.txt
openai
google-generativeai
beautifulsoup4
selenium